imu_orientation/visualizer/requirements.txt
pyserial>=3.5
pygame>=2.5.0
PyOpenGL>=3.1.7
